+++ packages/pn-pa-webapp/src/redux/newNotification/reducers.ts
@@ -7,6 +7,7 @@ import {
   NewNotificationRecipient,
   NotificationFeePolicy,
   PagoPaIntegrationMode,
+  PaymentModel,
   PreliminaryInformationsPayload,
 } from '../../models/NewNotification';
 import { UserGroup } from '../../models/user';
@@ -113,13 +114,23 @@ const newNotificationSlice = createSlice({
           payments: updatedPayments,
         };
       });
+
+      if (recipients.every((r) => r.debtPosition === PaymentModel.NOTHING)) {
+        state.notification = {
+          ...state.notification,
+          notificationFeePolicy: '' as NotificationFeePolicy,
+          paFee: undefined,
+          vat: undefined,
+          pagoPaIntMode: undefined,
+        };
+      }
     },
     setDebtPositionDetail: (
       state,
       action: PayloadAction<{
         recipients: Array<NewNotificationRecipient>;
         vat?: number;
-        paFee?: number;
+        paFee?: string;
         notificationFeePolicy: NotificationFeePolicy;
         pagoPaIntMode?: PagoPaIntegrationMode;
       }>
@@ -128,7 +139,7 @@ const newNotificationSlice = createSlice({
         ...state.notification,
         recipients: action.payload.recipients,
         vat: action.payload.vat,
-        paFee: action.payload.paFee && Number(action.payload.paFee),
+        paFee: action.payload.paFee,
         notificationFeePolicy: action.payload.notificationFeePolicy,
         pagoPaIntMode: action.payload.pagoPaIntMode,
       };