+++ packages/pn-personafisica-webapp/src/components/Contacts/EmailSmsContactWizard.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { Stack, Typography } from '@mui/material';
+import { Divider, Stack, Typography } from '@mui/material';
 import { appStateActions } from '@pagopa-pn/pn-commons';
 
 import { PFEventsType } from '../../models/PFEventsType';
@@ -14,6 +14,7 @@ import {
 import { createOrUpdateAddress } from '../../redux/contact/actions';
 import { contactsSelectors } from '../../redux/contact/reducers';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
+import { RootState } from '../../redux/store';
 import PFEventStrategyFactory from '../../utility/MixpanelUtils/PFEventStrategyFactory';
 import { contactAlreadyExists, internationalPhonePrefix } from '../../utility/contacts.utility';
 import ContactCodeDialog from './ContactCodeDialog';
@@ -33,7 +34,12 @@ const EmailSmsContactWizard: React.FC = () => {
   const { defaultSMSAddress, defaultEMAILAddress, addresses } = useAppSelector(
     contactsSelectors.selectAddresses
   );
-  const digitalContactRef = useRef<{ toggleEdit: () => void; resetForm: () => Promise<void> }>({
+  const externalEvent = useAppSelector((state: RootState) => state.contactsState.event);
+  const emailContactRef = useRef<{ toggleEdit: () => void; resetForm: () => Promise<void> }>({
+    toggleEdit: () => {},
+    resetForm: () => Promise.resolve(),
+  });
+  const smsContactRef = useRef<{ toggleEdit: () => void; resetForm: () => Promise<void> }>({
     toggleEdit: () => {},
     resetForm: () => Promise.resolve(),
   });
@@ -48,13 +54,14 @@ const EmailSmsContactWizard: React.FC = () => {
   const smsValue = defaultSMSAddress?.value ?? '';
 
   const handleSubmit = (channelType: ChannelType, value: string) => {
+    const source = externalEvent?.source ?? ContactSource.RECAPITI;
     PFEventStrategyFactory.triggerEvent(
       channelType === ChannelType.EMAIL
         ? PFEventsType.SEND_ADD_EMAIL_START
         : PFEventsType.SEND_ADD_SMS_START,
       {
         senderId: 'default',
-        source: ContactSource.RECAPITI,
+        source,
       }
     );
     // eslint-disable-next-line functional/immutable-data
@@ -69,7 +76,12 @@ const EmailSmsContactWizard: React.FC = () => {
 
   const handleCodeVerification = (channelType: ChannelType, verificationCode?: string) => {
     if (verificationCode) {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_UX_CONVERSION, 'default');
+      PFEventStrategyFactory.triggerEvent(
+        channelType === ChannelType.EMAIL
+          ? PFEventsType.SEND_ADD_EMAIL_UX_CONVERSION
+          : PFEventsType.SEND_ADD_SMS_UX_CONVERSION,
+        'default'
+      );
     }
 
     const digitalAddressParams: SaveDigitalAddressParams = {
@@ -98,7 +110,7 @@ const EmailSmsContactWizard: React.FC = () => {
           channelType === ChannelType.EMAIL
             ? PFEventsType.SEND_ADD_EMAIL_UX_SUCCESS
             : PFEventsType.SEND_ADD_SMS_UX_SUCCESS,
-          'default'
+          { senderId: 'default', fromSercqSend: true }
         );
 
         // contact has already been verified
@@ -106,14 +118,17 @@ const EmailSmsContactWizard: React.FC = () => {
         dispatch(
           appStateActions.addSuccess({
             title: '',
-            message: t(`courtesy-contacts.email-added-successfully`, {
+            message: t(`courtesy-contacts.${channelType.toLowerCase()}-added-successfully`, {
               ns: 'recapiti',
             }),
           })
         );
         setModalOpen(null);
-        if (emailValue || smsValue) {
-          digitalContactRef.current.toggleEdit();
+        if (currentAddress.current.channelType === ChannelType.EMAIL && emailValue) {
+          emailContactRef.current.toggleEdit();
+        }
+        if (currentAddress.current.channelType === ChannelType.SMS && smsValue) {
+          smsContactRef.current.toggleEdit();
         }
       })
       .catch(() => {});
@@ -121,10 +136,13 @@ const EmailSmsContactWizard: React.FC = () => {
 
   const handleCancelCode = async () => {
     setModalOpen(null);
-    if (emailValue || smsValue) {
-      digitalContactRef.current.toggleEdit();
+    if (currentAddress.current.channelType === ChannelType.EMAIL && emailValue) {
+      emailContactRef.current.toggleEdit();
+      await emailContactRef.current.resetForm();
+    } else if (currentAddress.current.channelType === ChannelType.SMS && smsValue) {
+      smsContactRef.current.toggleEdit();
+      await smsContactRef.current.resetForm();
     }
-    await digitalContactRef.current.resetForm();
   };
 
   return (
@@ -142,7 +160,7 @@ const EmailSmsContactWizard: React.FC = () => {
         label={t(`courtesy-contacts.email-to-add`, { ns: 'recapiti' })}
         value={emailValue}
         channelType={ChannelType.EMAIL}
-        ref={digitalContactRef}
+        ref={emailContactRef}
         inputProps={{
           label: t(`courtesy-contacts.link-email-placeholder`, {
             ns: 'recapiti',
@@ -159,35 +177,44 @@ const EmailSmsContactWizard: React.FC = () => {
           button: {
             sx: { height: '43px', fontWeight: 700, flexBasis: { xs: 'unset', lg: '25%' } },
           },
+          container: {
+            width: '100%',
+          },
         }}
       />
 
       {/* SMS */}
       {smsValue ? (
-        <DigitalContact
-          label={t(`courtesy-contacts.sms-to-add`, { ns: 'recapiti' })}
-          value={smsValue}
-          channelType={ChannelType.SMS}
-          ref={digitalContactRef}
-          inputProps={{
-            label: t(`courtesy-contacts.link-sms-placeholder`, {
-              ns: 'recapiti',
-            }),
-            prefix: internationalPhonePrefix,
-          }}
-          insertButtonLabel={t(`courtesy-contacts.sms-add`, { ns: 'recapiti' })}
-          onSubmit={(value) => handleSubmit(ChannelType.SMS, value)}
-          showVerifiedIcon
-          showLabelOnEdit
-          slotsProps={{
-            textField: {
-              sx: { flexBasis: { xs: 'unset', lg: '50%' } },
-            },
-            button: {
-              sx: { height: '43px', fontWeight: 700, flexBasis: { xs: 'unset', lg: '25%' } },
-            },
-          }}
-        />
+        <>
+          <Divider sx={{ mt: 1, mb: 3 }} />
+          <DigitalContact
+            label={t(`courtesy-contacts.sms-to-add`, { ns: 'recapiti' })}
+            value={smsValue}
+            channelType={ChannelType.SMS}
+            ref={smsContactRef}
+            inputProps={{
+              label: t(`courtesy-contacts.link-sms-placeholder`, {
+                ns: 'recapiti',
+              }),
+              prefix: internationalPhonePrefix,
+            }}
+            insertButtonLabel={t(`courtesy-contacts.sms-add`, { ns: 'recapiti' })}
+            onSubmit={(value) => handleSubmit(ChannelType.SMS, value)}
+            showVerifiedIcon
+            showLabelOnEdit
+            slotsProps={{
+              textField: {
+                sx: { flexBasis: { xs: 'unset', lg: '50%' } },
+              },
+              button: {
+                sx: { height: '43px', fontWeight: 700, flexBasis: { xs: 'unset', lg: '25%' } },
+              },
+              container: {
+                width: '100%',
+              },
+            }}
+          />
+        </>
       ) : (
         <SmsContactItem
           slotsProps={{
@@ -208,7 +235,13 @@ const EmailSmsContactWizard: React.FC = () => {
         open={modalOpen === ModalType.CODE}
         onConfirm={(code) => handleCodeVerification(currentAddress.current.channelType, code)}
         onDiscard={handleCancelCode}
-        onError={() => PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_CODE_ERROR)}
+        onError={() =>
+          PFEventStrategyFactory.triggerEvent(
+            currentAddress.current.channelType === ChannelType.EMAIL
+              ? PFEventsType.SEND_ADD_EMAIL_CODE_ERROR
+              : PFEventsType.SEND_ADD_SMS_CODE_ERROR
+          )
+        }
       />
       <ExistingContactDialog
         open={modalOpen === ModalType.EXISTING}